+++ src/lib.rs
@@ -86,7 +86,9 @@ pub use crate::algorithms::paths::radii::{
 };
 pub use crate::algorithms::paths::random_walk::random_walk;
 pub use crate::algorithms::paths::widest_path::{
-    widest_path, widest_path_widths, widest_path_widths_with_mode, widest_path_with_mode,
+    widest_path, widest_path_widths, widest_path_widths_floyd_warshall,
+    widest_path_widths_floyd_warshall_with_mode, widest_path_widths_with_mode,
+    widest_path_with_mode,
 };
 pub use crate::algorithms::properties::assortativity::{
     assortativity_degree, assortativity_degree_directed,